rename networkstore to portstore to match reality
Follow-up to PR 4290
